shared 페이지 분리
