+++ part2/src/pages/FolderPage.js
@@ -1,0 +1,5 @@
+function FolderPage() {
+  return <></>;
+}
+
+export default FolderPage;
